requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-aiplatform
google-cloud-texttospeech
psutil
numpy
requests
python-dotenv
sounddevice
wavio
scipy